Create Pulitzer Winners
The purpose of this project was to create two tables and practice joining them in meaningful ways. The first join included data from from table one of author name and current city of residence along with their debut novel title and publication year. In the second join, I wanted to join the data so it was easy to compare what length of time there was between the debut novel and the Pulitzer winning novel.