Create CityBikeLane
The main focus I had when manipulating this data set was determining a way to group ratings by street name, finding an average of those ratings, and then using that data to make recommendations to either leave the bike lane "as is", to remove certain bike lanes due to safety,  or the recommend improvements. I partitioned the data by street to get an average of all inspection ratings to make the most informed decision.